Add a ThemeRoller-compatible UI theme.
This is the first version of what is intended to be the "default" UI theme for scripty2 UI controls. This is a happy medium between:

1. not prescribing look and feel _at all_ (poor out-of-box experience for devs);
2. prescribing look and feel down to the pixel (too opinionated and rigid).

Ideally, I'd like to have several "defaults" with different color schemes: one with a black base instead of white, several hue alternatives in case the green clashes with the rest of your site, etc.

This theme is also compatible(-ish) with jQuery UI components, with eventual goal of 100% compatibility. And, yes, it's got a long way to go.