Defer placement of autocomplete menu.
In IE the layout of the input element sometimes is not determinable if the
autocomplete widget is being initialized early in the page loading process.
The sizes specs coming back from IE are all 0.

There might be a more elegant/robust solution to this problem but simply
surrounded the positioning with an anonymous defer seems to solve the problem
in my app.