Check for selected item when setting the input.
Sometimes (especially noticed in IE) the menu will fire a selected event
but the element will be undefined. Tracing through the code I determined the
index in the menu is being set to -1.

I thought about not sending the event if the index is -1 but thought there
might be some other code using a menu that would want to know even in the
case of -1. So instead I just had the autocomplete code check for an element
and only set the input if an element is provided by the event.